+++ src/main/java/roomescape/admin/presentation/controller/AdminViewController.java
@@ -1,4 +1,4 @@
-package roomescape.admin.presentation;
+package roomescape.admin.presentation.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
