+++ src/main/java/roomescape/auth/presentation/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package roomescape.member.infrastructure;
+package roomescape.auth.presentation;
 
 import jakarta.servlet.http.HttpServletRequest;
