+++ src/main/java/roomescape/auth/presentation/controller/AuthController.java
@@ -1,35 +1,32 @@
-package roomescape.member.presentation;
+package roomescape.auth.presentation.controller;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.member.application.dto.GetMemberResponse;
-import roomescape.member.application.service.MemberService;
+import roomescape.auth.application.service.AuthService;
+import roomescape.auth.presentation.dto.LoginRequest;
+import roomescape.auth.presentation.dto.TokenResponse;
 import roomescape.member.domain.Member;
-import roomescape.member.presentation.dto.LoginRequest;
 import roomescape.member.presentation.dto.MemberNameResponse;
-import roomescape.member.presentation.dto.RegisterRequest;
-import roomescape.member.presentation.dto.TokenResponse;
 
 @RestController
-public class MemberController {
+public class AuthController {
 
     private static final String SET_COOKIE_KEY = "token";
 
-    private final MemberService memberService;
+    private final AuthService authService;
 
-    public MemberController(MemberService memberService) {
-        this.memberService = memberService;
+    public AuthController(AuthService authService) {
+        this.authService = authService;
     }
 
     @PostMapping("/login")
     public ResponseEntity<Void> login(@RequestBody LoginRequest loginRequest, HttpServletResponse response) {
-        TokenResponse token = memberService.login(loginRequest);
+        TokenResponse token = authService.login(loginRequest);
         Cookie cookie = new Cookie(SET_COOKIE_KEY, token.accessToken());
         cookie.setHttpOnly(true);
         cookie.setPath("/");
@@ -48,17 +45,5 @@ public class MemberController {
         logoutCookie.setMaxAge(0);
         response.addCookie(logoutCookie);
         return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/members")
-    public ResponseEntity<MemberNameResponse> register(@RequestBody RegisterRequest registerRequest) {
-        MemberNameResponse response = memberService.signup(registerRequest);
-        return ResponseEntity.status(201).body(response);
-    }
-
-    @GetMapping("/members")
-    public ResponseEntity<List<GetMemberResponse>> getMembers() {
-        List<GetMemberResponse> members = memberService.getMembers();
-        return ResponseEntity.ok().body(members);
     }
 }
