+++ src/main/java/roomescape/auth/presentation/controller/AuthViewController.java
@@ -1,16 +1,10 @@
-package roomescape.member.presentation;
+package roomescape.auth.presentation.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class MemberViewController {
-    @GetMapping("/reservation")
-    public String reservation(
-    ) {
-        return "/reservation";
-    }
-
+public class AuthViewController {
     @GetMapping("/login")
     public String login(
     ) {
