+++ src/main/java/roomescape/auth/presentation/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member.presentation.dto;
+package roomescape.auth.presentation.dto;
 
 public record TokenResponse(String accessToken) {
 }
