+++ src/test/java/roomescape/global/ApiHelper.java
@@ -4,8 +4,8 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
+import roomescape.auth.presentation.dto.LoginRequest;
 import roomescape.member.domain.Role;
-import roomescape.member.presentation.dto.LoginRequest;
 import roomescape.member.presentation.dto.RegisterRequest;
 import roomescape.member.presentation.fixture.MemberFixture;
