+++ src/test/java/roomescape/member/application/service/MemberServiceTest.java
@@ -1,20 +1,14 @@
 package roomescape.member.application.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import roomescape.global.exception.InvalidMemberException;
 import roomescape.member.application.dto.GetMemberResponse;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
-import roomescape.member.infrastructure.JwtTokenProvider;
 import roomescape.member.infrastructure.fake.FakeMemberDao;
-import roomescape.member.presentation.dto.LoginRequest;
-import roomescape.member.presentation.dto.TokenResponse;
 
 class MemberServiceTest {
 
@@ -23,56 +17,7 @@ class MemberServiceTest {
 
     MemberServiceTest() {
         this.userRepository = new FakeMemberDao();
-        this.memberService = new MemberService(new JwtTokenProvider("secret-key", 360000), userRepository);
-    }
-
-    @Test
-    @DisplayName("로그인 테스트")
-    void loginTest() {
-        // given
-        String email = "email@email.com";
-        String password = "password";
-        userRepository.insert(new Member(0L, "name", email, password, Role.USER));
-        LoginRequest loginRequest = new LoginRequest(email, password);
-
-        // when
-        TokenResponse token = memberService.login(loginRequest);
-
-        // then
-        assertThat(token.accessToken()).isNotEmpty();
-    }
-
-    @Test
-    @DisplayName("비밀번호가 틀릴 경우 로그인에 실패한다.")
-    void loginFailedTest() {
-        // given
-        String email = "email@email.com";
-        String password = "password";
-        userRepository.insert(new Member(0L, "name", email, password, Role.USER));
-        LoginRequest loginRequest = new LoginRequest(email, "wrong password");
-
-        // when - then
-        assertThatThrownBy(() -> memberService.login(loginRequest))
-                .isInstanceOf(InvalidMemberException.class)
-                .hasMessage("비밀번호가 틀렸습니다.")
-                .hasFieldOrPropertyWithValue("statusCode", HttpStatus.UNAUTHORIZED);
-    }
-
-    @Test
-    @DisplayName("사용자 정보 조회 테스트")
-    void getMemberTest() {
-        // given
-        String email = "email@email.com";
-        String password = "password";
-        userRepository.insert(new Member(0L, "name", email, password, Role.USER));
-        LoginRequest loginRequest = new LoginRequest(email, password);
-        TokenResponse token = memberService.login(loginRequest);
-
-        // when
-        Member member = memberService.getMember(token.accessToken());
-
-        // then
-        assertThat(member.getName()).isEqualTo("name");
+        this.memberService = new MemberService(userRepository);
     }
 
     @Test
